move all models to models.py
