+++ src/api/modules/generators.py
@@ -2,27 +2,12 @@
 
 import uuid
 from random import randint
-from typing import Literal
 
 from fastapi import APIRouter
 
-# pylint: disable-next=no-name-in-module
-from pydantic import BaseModel, Field, validator
+from api.models import Numbers, NumbersConfig, UUIDConfig, UUIDs
 
 router = APIRouter(prefix="/generators", tags=["generators"])
-
-
-class UUIDConfig(BaseModel):
-    """Model to hold configuration for UUID generation"""
-
-    uuid_type: Literal[1, 4]
-    quantity: int = Field(gt=0, default=1)
-
-
-class UUIDs(BaseModel):
-    """Model to hold a list of UUIDs"""
-
-    uuids: list[str]
 
 
 @router.post("/uuids/")
@@ -36,28 +21,6 @@ async def bulk_uuids(config: UUIDConfig) -> UUIDs:
         raise ValueError(f"unsupported UUID type: {config.uuid_type}")
     uuids = [str(function()) for _ in range(config.quantity)]
     return UUIDs(uuids=uuids)
-
-
-class NumbersConfig(BaseModel):
-    """Model to hold configuration for number generation"""
-
-    lower_bound: int = Field(default=1)
-    upper_bound: int = Field(default=1)
-    quantity: int = Field(gt=0, default=1)
-
-    @validator("upper_bound")
-    def upper_bound_must_be_greater_than_lower_bound(cls, v, values):
-        """Confirm upper bound is greater than lower bound"""
-        if "lower_bound" in values and v < values["lower_bound"]:
-            raise ValueError("upper bound must be greater than lower bound")
-        return v
-
-
-class Numbers(BaseModel):
-    """Model to hold a list of numbers"""
-
-    numbers: list[int]
-    total: int
 
 
 @router.post("/random-numbers/")
